line.c angepasst, Tasterabfrage implementiert
bugfixes:
- get_line_adc(void): korrigiert, lieferte vorher 8bit Wert statt 16bit
- get_line_adc(void): Warteschleife vor dem Kanalwechsel des ADC eingefügt, da sonst der Kanalwechsel bei zwei direkt aufeinanderfolgenden ADC Messungen noch gesperrt ist und somit der vorherige Kanal nochmals gemessen wird

new features:
- Taster interrupt

faults:
- Erkennung der einzelnen Taster funktioniert nicht korrekt

+++ Asuro/line.h
@@ -31,7 +31,7 @@ void init_line(void);
 *           65536   max. hell
 *
 ************************************************************************/
-uint8_t get_line_adc(char side);
+uint16_t get_line_adc(char side);
 
 
 #endif /* LINE_H_ */
